refactor(sdk): share reward protocol result models
