+++ rock/sdk/reward/__init__.py
@@ -1,0 +1,21 @@
+"""Shared reward protocol models."""
+
+from rock.sdk.reward.result import (
+    AgentInfo,
+    AgentResult,
+    ExceptionInfo,
+    ModelInfo,
+    RewardTrialResult,
+    TimingInfo,
+    VerifierResult,
+)
+
+__all__ = [
+    "AgentInfo",
+    "AgentResult",
+    "ExceptionInfo",
+    "ModelInfo",
+    "RewardTrialResult",
+    "TimingInfo",
+    "VerifierResult",
+]
